+++ src/pll.rs
@@ -4,17 +4,19 @@ use serde::{Deserialize, Serialize};
 /// Type-II, sampled phase, discrete time PLL
 ///
 /// This PLL tracks the frequency and phase of an input signal with respect to the sampling clock.
-/// The transfer function is I^2,I from input phase to output phase and P,I from input phase to
-/// output frequency.
+/// The open loop transfer function is I^2,I from input phase to output phase and P,I from input
+/// phase to output frequency.
+///
+/// The transfer functions (for phase and frequency) contain an additional zero at Nyquist.
 ///
 /// The PLL locks to any frequency (i.e. it locks to the alias in the first Nyquist zone) and is
 /// stable for any gain (1 <= shift <= 30). It has a single parameter that determines the loop
 /// bandwidth in octave steps. The gain can be changed freely between updates.
 ///
 /// The frequency and phase settling time constants for a frequency/phase jump are `1 << shift`
-/// update cycles. The loop bandwidth is about `1/(2*pi*(1 << shift))` in units of the sample rate.
-/// While the phase is being settled within one turn, there is a typically very small frequency
-/// overshoot.
+/// update cycles. The loop bandwidth is `1/(2*pi*(1 << shift))` in units of the sample rate.
+/// While the phase is being settled after settling the frequency, there is a typically very
+/// small frequency overshoot.
 ///
 /// All math is naturally wrapping 32 bit integer. Phase and frequency are understood modulo that
 /// overflow in the first Nyquist zone. Expressing the IIR equations in other ways (e.g. single
@@ -25,9 +27,10 @@ use serde::{Deserialize, Serialize};
 /// bias is applied. Rounding is "half up". The phase truncation error can be removed very
 /// efficiently by dithering.
 ///
-/// This PLL does not unwrap phase slips during lock acquisition. This can and should be
-/// implemented elsewhere by unwrapping and scaling the input phase and un-scaling
-/// and wrapping output phase and frequency. This affects dynamic range, gain, and noise accordingly.
+/// This PLL does not unwrap phase slips accumulated during (frequency) lock acquisition.
+/// This can and should be implemented elsewhere by unwrapping and scaling the input phase
+/// and un-scaling and wrapping output phase and frequency. This then affects dynamic range,
+/// gain, and noise accordingly.
 ///
 /// The extension to I^3,I^2,I behavior to track chirps phase-accurately or to i64 data to
 /// increase resolution for extremely narrowband applications is obvious.
@@ -53,31 +56,44 @@ impl PLL {
     ///   per update. A good value is typically `shift_frequency - 1`.
     ///
     /// Returns:
-    /// A tuple of instantaneous phase and frequency (the current phase increment).
+    /// A tuple of instantaneous phase and frequency estimates.
     pub fn update(&mut self, x: Option<i32>, shift_frequency: u32, shift_phase: u32) -> (i32, i32) {
         debug_assert!((1..=30).contains(&shift_frequency));
         debug_assert!((1..=30).contains(&shift_phase));
-        let f = if let Some(x) = x {
-            let e = x.wrapping_sub(self.f);
-            self.f = self.f.wrapping_add(unchecked_shr(
-                unchecked_shl(1i32, shift_frequency - 1)
-                    .wrapping_add(e)
-                    .wrapping_sub(self.x),
-                shift_frequency,
-            ));
+        if let Some(x) = x {
+            let df = unchecked_shr(
+                unchecked_shl(1, shift_frequency - 1)
+                .wrapping_add(x)
+                .wrapping_sub(self.x)
+                .wrapping_sub(self.f),
+                shift_frequency);
             self.x = x;
-            self.f.wrapping_add(unchecked_shr(
-                unchecked_shl(1i32, shift_phase - 1)
-                    .wrapping_add(e)
-                    .wrapping_sub(self.y),
-                shift_phase,
-            ))
+            self.f = self.f.wrapping_add(df);
+            let f = self.f.wrapping_sub(df >> 1);
+            self.y = self.y.wrapping_add(f);
+            let dy = unchecked_shr(
+                unchecked_shl(1, shift_phase - 1)
+                .wrapping_add(x)
+                .wrapping_sub(self.y),
+                shift_phase);
+            self.y = self.y.wrapping_add(dy);
+            let y = self.y.wrapping_sub(dy >> 1);
+            (y, f.wrapping_add(dy))
         } else {
             self.x = self.x.wrapping_add(self.f);
-            self.f
-        };
-        self.y = self.y.wrapping_add(f);
-        (self.y, f)
+            self.y = self.y.wrapping_add(self.f);
+            (self.y, self.f)
+        }
+    }
+
+    /// Return the current phase estimate
+    pub fn phase(&self) -> i32 {
+        self.y
+    }
+
+    /// Return the current frequency estimate
+    pub fn frequency(&self) -> i32 {
+        self.f
     }
 }
 
@@ -88,8 +104,8 @@ mod tests {
     fn mini() {
         let mut p = PLL::default();
         let (y, f) = p.update(Some(0x10000), 8, 4);
-        assert_eq!(y, 0x1100);
-        assert_eq!(f, y);
+        assert_eq!(y, 0x87c);
+        assert_eq!(f, 0x1078);
     }
 
     #[test]
@@ -103,7 +119,8 @@ mod tests {
             x = x.wrapping_add(f0);
             let (y, f) = p.update(Some(x), shift.0, shift.1);
             if i > n / 4 {
-                assert_eq!(f.wrapping_sub(f0).abs() <= 1, true);
+                // The remaining error would be removed by dithering.
+                assert_eq!(f.wrapping_sub(f0).abs() <= 1 << 10, true);
             }
             if i > n / 2 {
                 // The remaining error would be removed by dithering.